0907 Django

+++ Django/05_django/articles/views.py
@@ -1,72 +1,49 @@
 from django.shortcuts import render, redirect
-from django.views.decorators.http import require_safe, require_http_methods, require_POST
+from django.views.decorators.http import require_http_methods, require_POST, require_safe
+from django.contrib.auth.decorators import login_required
 from .models import Article
 from .forms import ArticleForm
-
 
 # Create your views here.
 @require_safe
 def index(request):
-    # DB에 전체 데이터를 조회
     articles = Article.objects.all()
     context = {
         'articles': articles,
     }
     return render(request, 'articles/index.html', context)
 
-
-# def new(request):
-#     form = ArticleForm()
-#     context = {
-#         'form': form,
-#     }
-#     return render(request, 'articles/new.html', context)
-
-
+@login_required
 @require_http_methods(['GET', 'POST'])
 def create(request):
     if request.method == 'POST':
-        # create
         form = ArticleForm(request.POST)
         if form.is_valid():
             article = form.save()
             return redirect('articles:detail', article.pk)
     else:
-        # new
         form = ArticleForm()
     context = {
         'form': form,
     }
     return render(request, 'articles/create.html', context)
 
-
 @require_safe
 def detail(request, pk):
-    # variable routing으로 받은 pk 값으로 데이터를 조회
     article = Article.objects.get(pk=pk)
     context = {
         'article': article,
     }
     return render(request, 'articles/detail.html', context)
 
-
 @require_POST
 def delete(request, pk):
-    article = Article.objects.get(pk=pk)
-    article.delete()
+    if request.user.is_authenticated:
+        article = Article.objects.get(pk=pk)
+        article.delete()
     return redirect('articles:index')
 
-
-# def edit(request, pk):
-#     article = Article.objects.get(pk=pk)
-#     form = ArticleForm(instance=article)
-#     context = {
-#         'article': article,
-#         'form': form,
-#     }
-#     return render(request, 'articles/edit.html', context)
-
-
+@login_required
 @require_http_methods(['GET', 'POST'])
 def update(request, pk):
     article = Article.objects.get(pk=pk)
@@ -79,7 +56,7 @@ def update(request, pk):
     else:
         form = ArticleForm(instance=article)
     context = {
-        'article': article,
         'form': form,
+        'article': article,
     }
     return render(request, 'articles/update.html', context)
